GROK-14119: ODEs: Refactoring
Code refactoring with respect to the code review
