+++ packages/ODEs/CHANGELOG.md
@@ -1,4 +1,4 @@
-# ODEs changelog
+# Deqs changelog
 
 ## 1.0.0 (WIP)
