+++ packages/ODEs/package.json
@@ -1,7 +1,7 @@
 {
-	"name": "@datagrok/odes",
-	"friendlyName": "ODEs",
-	"version": "0.0.1",
+	"name": "@datagrok/deqs",
+	"friendlyName": "Deqs",
+	"version": "1.0.0",
 	"description": "WebAutoSolver tools for solving ODEs in a browser.",
 	"author": {
 		"name": "Viktor Makarichev",
